requirements.txt
Django==5.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
Pillow==11.1.0
python-decouple==3.8
channels==4.2.0
channels-redis==4.4.0
gunicorn==21.2.0
whitenoise==6.6.0
